refactor: 패키지구조 수정 및 린트정리
