+++ app/src/main/java/com/team/recordream/di/RepositoryModule.kt
@@ -43,7 +43,7 @@ object RepositoryModule {
     @Provides
     @Singleton
     fun providesDocumentRepository(
-        documentService: DocumentService
+        documentService: DocumentService,
     ): DocumentRepository = DocumentRepositoryImpl(documentService)
 
     @Provides
