+++ app/src/main/java/com/team/recordream/presentation/MainActivity.kt
@@ -7,10 +7,10 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
 import com.team.recordream.R
-import com.team.recordream.base.BindingActivity
 import com.team.recordream.databinding.ActivityMainBinding
 import com.team.recordream.presentation.MainActivity.FragmentType.HOME
 import com.team.recordream.presentation.MainActivity.FragmentType.STORAGE
+import com.team.recordream.presentation.common.BindingActivity
 import com.team.recordream.presentation.home.HomeFragment
 import com.team.recordream.presentation.mypage.MypageActivity
 import com.team.recordream.presentation.record.RecordActivity
