+++ app/src/main/java/com/team/recordream/presentation/common/BindingFragment.kt
@@ -1,4 +1,4 @@
-package com.team.recordream.base
+package com.team.recordream.presentation.common
 
 import android.os.Bundle
 import android.view.LayoutInflater
