+++ app/src/main/java/com/team/recordream/presentation/common/model/PlayButtonState.kt
@@ -1,0 +1,6 @@
+package com.team.recordream.presentation.common.model // ktlint-disable package-name
+
+enum class PlayButtonState {
+    RECORDER_PLAY,
+    RECORDER_STOP,
+}
