--- app/src/main/java/com/team/recordream/presentation/record/recording/uistate/PlayButtonState.kt
@@ -1,6 +1,0 @@
-package com.team.recordream.presentation.record.recording.uistate // ktlint-disable package-name
-
-enum class PlayButtonState {
-    RECORDER_PLAY,
-    RECORDER_STOP,
-}
