+++ app/src/main/java/com/team/recordream/presentation/splash/SplashActivity.kt
@@ -8,9 +8,9 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.team.recordream.R
-import com.team.recordream.base.BindingActivity
 import com.team.recordream.databinding.ActivitySplashBinding
 import com.team.recordream.presentation.MainActivity
+import com.team.recordream.presentation.common.BindingActivity
 import com.team.recordream.presentation.login.LoginActivity
 import com.team.recordream.presentation.splash.SplashViewModel.LoginState
 import dagger.hilt.android.AndroidEntryPoint
