+++ app/src/main/java/com/team/recordream/util/BindingAdapter.kt
@@ -6,10 +6,10 @@ import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
 import com.google.android.material.chip.Chip
 import com.team.recordream.R
+import com.team.recordream.presentation.common.model.PlayButtonState
+import com.team.recordream.presentation.common.model.PlayButtonState.*
 import com.team.recordream.presentation.record.model.Emotion
 import com.team.recordream.presentation.record.model.Emotion.*
-import com.team.recordream.presentation.record.recording.uistate.PlayButtonState
-import com.team.recordream.presentation.record.recording.uistate.PlayButtonState.*
 import com.team.recordream.presentation.record.recording.uistate.RecordButtonState
 import com.team.recordream.presentation.record.recording.uistate.RecordButtonState.*
