+++ app/src/main/java/com/team/recordream/util/ZoomOutPageTransformer.kt
@@ -34,9 +34,9 @@ class ZoomOutPageTransformer : ViewPager2.PageTransformer {
                     scaleY = scaleFactor
                     // Fade the page relative to its size.
                     alpha = (
-                            MIN_ALPHA +
-                                    (((scaleFactor - MIN_SCALE) / (1 - MIN_SCALE)) * (1 - MIN_ALPHA))
-                            )
+                        MIN_ALPHA +
+                            (((scaleFactor - MIN_SCALE) / (1 - MIN_SCALE)) * (1 - MIN_ALPHA))
+                        )
                 }
 
                 else -> { // (1,+Infinity]
