+++ app/src/main/res/drawable/insta_card_pink.xml
@@ -3,11 +3,11 @@
     android:height="468dp"
     android:viewportWidth="300"
     android:viewportHeight="468">
-  <path
-      android:pathData="M25,0L275,0A25,25 0,0 1,300 25L300,443A25,25 0,0 1,275 468L25,468A25,25 0,0 1,0 443L0,25A25,25 0,0 1,25 0z"
-      android:fillColor="#02040F"/>
-  <path
-      android:pathData="M25,0L275,0A25,25 0,0 1,300 25L300,443A25,25 0,0 1,275 468L25,468A25,25 0,0 1,0 443L0,25A25,25 0,0 1,25 0z"
-      android:fillColor="#FFB6E6"
-      android:fillAlpha="0.07"/>
+    <path
+        android:fillColor="#02040F"
+        android:pathData="M25,0L275,0A25,25 0,0 1,300 25L300,443A25,25 0,0 1,275 468L25,468A25,25 0,0 1,0 443L0,25A25,25 0,0 1,25 0z" />
+    <path
+        android:fillAlpha="0.07"
+        android:fillColor="#FFB6E6"
+        android:pathData="M25,0L275,0A25,25 0,0 1,300 25L300,443A25,25 0,0 1,275 468L25,468A25,25 0,0 1,0 443L0,25A25,25 0,0 1,25 0z" />
 </vector>
