+++ app/src/main/res/layout/item_list_emotion.xml
@@ -34,9 +34,9 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
-            app:emotionText="@{emotion.selectedEmotion}"
             android:textAppearance="@style/R10"
             android:textColor="@color/selector_color_white01_white03"
+            app:emotionText="@{emotion.selectedEmotion}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="@+id/iv_record_item_emotion"
             app:layout_constraintStart_toStartOf="@+id/iv_record_item_emotion"
