+++ app/src/main/res/layout/item_list_store_grid.xml
@@ -12,8 +12,8 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_marginHorizontal="5dp"
         android:layout_height="wrap_content"
+        android:layout_marginHorizontal="5dp"
         android:paddingBottom="10dp">
 
         <androidx.constraintlayout.widget.ConstraintLayout
